deployment/staging/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.35.0
gunicorn==23.0.0
openai==1.102.0
chromadb==1.0.16
python-dotenv==1.1.1
pydantic==2.11.7
